pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "signet-pqc-control-plane"
version = "0.1.0"
description = "Signet PQC Control Plane (MVP): policy -> enforcement -> cryptographic proof."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "Your Name", email = "you@example.com" }
]
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn>=0.29.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.1",
  "orjson>=3.10.0",
  "PyNaCl>=1.5.0",
  "python-multipart>=0.0.9",
  "httpx>=0.27.0",
]

[project.scripts]
spcp_cli = "spcp.spcp_cli:main"
cbom-agent = "cbom_agent.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
  "types-requests",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E","F","I","B","UP","Q","S","PGH","PTH","C4"]
ignore = ["S101"]

[tool.pytest.ini_options]
addopts = "-q --maxfail=1 --disable-warnings --cov=spcp --cov-report=xml"
pythonpath = ["src"]
